[#168289372] Renamed the details of the services
Renamed the details of the services
